net: relocate socket headers under subdir
rsocket.h, rsocketaddress.h and riosocket.h are the socket
abstraction layer that the higher-level net/* (rresolve, rhttp,
rtls, rsrtp, ...) and ev/* (revtcp, revudp) headers build on. They
fit semantically with net/ and there is nothing else at the top
level that uses them — moving clarifies the boundary between
network primitives and truly foundational primitives.

Affects 14 #include sites plus the meson source list and the
private rsocket-private.h relative paths in rlib/ev. No public
API change beyond the include path.
