+++ include/rlib/ev/revtcp.h
@@ -26,8 +26,8 @@
 
 #include <rlib/ev/revloop.h>
 #include <rlib/rbuffer.h>
-#include <rlib/rsocketaddress.h>
-#include <rlib/rsocket.h>
+#include <rlib/net/rsocketaddress.h>
+#include <rlib/net/rsocket.h>
 #include <rlib/rref.h>
 
 R_BEGIN_DECLS
