+++ include/rlib/ev/revudp.h
@@ -26,7 +26,7 @@
 
 #include <rlib/ev/revloop.h>
 #include <rlib/rbuffer.h>
-#include <rlib/rsocketaddress.h>
+#include <rlib/net/rsocketaddress.h>
 #include <rlib/rref.h>
 
 R_BEGIN_DECLS
