+++ include/rlib/net/proto/rsdp.h
@@ -27,7 +27,7 @@
 
 #include <rlib/rbuffer.h>
 #include <rlib/crypto/rmsgdigest.h>
-#include <rlib/rsocketaddress.h>
+#include <rlib/net/rsocketaddress.h>
 #include <rlib/rstr.h>
 #include <rlib/ruri.h>
