+++ include/rlib/net/proto/rstun.h
@@ -24,7 +24,7 @@
 
 #include <rlib/rtypes.h>
 
-#include <rlib/rsocketaddress.h>
+#include <rlib/net/rsocketaddress.h>
 
 R_BEGIN_DECLS
