+++ include/rlib/net/riosocket.h
@@ -25,7 +25,7 @@
 #include <rlib/rtypes.h>
 
 #include <rlib/rbuffer.h>
-#include <rlib/rsocketaddress.h>
+#include <rlib/net/rsocketaddress.h>
 
 R_BEGIN_DECLS
