+++ include/rlib/net/rresolve.h
@@ -24,8 +24,8 @@
 
 #include <rlib/rtypes.h>
 
-#include <rlib/rsocket.h>
-#include <rlib/rsocketaddress.h>
+#include <rlib/net/rsocket.h>
+#include <rlib/net/rsocketaddress.h>
 
 R_BEGIN_DECLS
