+++ rlib/meson.build
@@ -10,7 +10,7 @@ rlib_sources = [
   'rcrc.c',
   'renv.c',
   'rio.c',
-  'riosocket.c',
+  'net/riosocket.c',
   'rjson.c',
   'rjsonparser.c',
   'rlibinit.c',
@@ -27,8 +27,8 @@ rlib_sources = [
   'rprng-mt.c',
   'rrand.c',
   'rref.c',
-  'rsocketaddress.c',
-  'rsocket.c',
+  'net/rsocketaddress.c',
+  'net/rsocket.c',
   'rstr.c',
   'rstrmatch.c',
   'rtaskqueue.c',
