+++ rlib/net/rsocket.c
